Fix most of the sheep tracking bugs in HnS
This is also present in the orignial version by DinnerBuffet. It appears
that in one point in time, the depth of the tiles shrunk by about 0.06
units.  The scripable play area zone was made to work with the original
tile height, with the shrink there is a 0.06 unit gap right above the
tiles, where small flat pieces (like sheep/wolf tokens that are 0.02
units deep) will fall into and not be tracked.

Moving down the scripting zone to accomidate completely fixes HnS.